added initial commands for adding friends
