+++ sendletters/SendLetters.csproj
@@ -66,10 +66,10 @@
     <Compile Include="Menus\ComposeLetter.cs" />
     <Compile Include="ModConfig.cs" />
     <Compile Include="Models\CreateMessageModel.cs" />
-    <Compile Include="SendLetterMod.cs" />
+    <Compile Include="ModEntry.cs" />
     <Compile Include="Domain\Player.cs" />
     <Compile Include="ModEvents.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="SendLetterMod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\ConfigurationService.cs" />
     <Compile Include="Services\IConfigurationService.cs" />
